requirements.txt
# fastapi>=0.100.0
# uvicorn==0.22.0
# python-multipart==0.0.6
# twilio==7.17.0
# python-dotenv>=0.21.0
# aiohttp>=3.8.1
# pydantic>=2.7.0
# langchain>=0.3.0
# openai>=1.6.1
# tiktoken>=0.5.2
# langchain-openai>=0.0.2
# langgraph>=0.0.10
# chromadb>=0.4.18
# sentence-transformers>=2.6.0
# beautifulsoup4>=4.12.2
# requests>=2.31.0
# typing-extensions>=4.8.0
# python-jose>=3.3.0
# langsmith>=0.1.112
# langchain-core>=0.3.0
# langchain-community>=0.3.0
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.9
# itsdangerous==2.2.0
# langchain-huggingface == 0.1.2
# cryptography== 44.0.0
# nltk==3.9.1
# textblob==0.19.0
# pandas==2.2.3
# # spacy==3.8.4
# python-multipart==0.0.6
# psycopg2-binary==2.9.9
# # pydantic==2.5.2
# prometheus-client==0.19.0  # For metrics
# sentry-sdk==1.32.0  # For error tracking
# assemblyai==0.36.0
# pydub==0.25.1
# python-decouple==3.8
# ffmpeg==1.4
# sendgrid==6.11.0
# aiosmtplib== 4.0.0
# loguru==0.7.3
# jinja2

fastapi>=0.100.0
uvicorn==0.22.0
python-multipart==0.0.6
twilio==7.17.0
python-dotenv>=0.21.0
aiohttp>=3.8.1
pydantic>=2.7.0
langchain>=0.3.0
openai>=1.6.1
tiktoken>=0.5.2
langchain-openai>=0.0.2
langgraph==0.2.69
langgraph-checkpoint==2.0.10
langgraph-checkpoint-postgres==2.0.13
beautifulsoup4>=4.12.2
requests>=2.31.0
typing-extensions>=4.8.0
python-jose>=3.3.0
langsmith>=0.1.112
langchain-core>=0.3.0
langchain-community>=0.3.0
sqlalchemy>=2.0.0
itsdangerous==2.2.0
cryptography== 44.0.0
textblob==0.19.0
pandas==2.2.3
# spacy==3.8.4
python-multipart==0.0.6
# pydantic==2.5.2
prometheus-client==0.19.0  # For metrics
sentry-sdk==1.32.0  # For error tracking
assemblyai==0.36.0
pydub==0.25.1
python-decouple==3.8
ffmpeg==1.4
sendgrid==6.11.0
aiosmtplib== 4.0.0
loguru==0.7.3
googletrans==4.0.2
httpx>=0.23.0
langchain-redis==0.1.2
unstructured
unstructured[docx]
boto3==1.36.19
jinja2
psycopg==3.2.4
psycopg-pool==3.2.4
psycopg[binary,pool]
PyPDF2==3.0.1
weasyprint==64.0
psycopg2-binary
google-cloud-speech==2.31.0
langchain-google-community